+++ plugins/Organic/Organic.h
@@ -78,7 +78,7 @@ const QString WAVEFORM_NAMES[6] = {
 	"Exponential wave"
 	};
 	
-const float CENT = 1.0f / 1200.0f;
+const double CENT = 1.0 / 1200.0;
 
 class OscillatorObject : public Model
 {
@@ -95,11 +95,11 @@ private:
 	float m_volumeLeft;
 	float m_volumeRight;
 	// normalized detuning -> x/sampleRate
-	float m_detuningLeft;
-	float m_detuningRight;
+	double m_detuningLeft;
+	double m_detuningRight;
 	// normalized offset -> x/360
-	float m_phaseOffsetLeft;
-	float m_phaseOffsetRight;
+	double m_phaseOffsetLeft;
+	double m_phaseOffsetRight;
 
 	OscillatorObject( Model * _parent, int _index );
 	~OscillatorObject() override = default;
@@ -133,7 +133,7 @@ public:
 
 	QString nodeName() const override;
 
-	static float * s_harmonics;
+	static double * s_harmonics;
 
 public slots:
 	void randomiseSettings();
